requirements.txt
numpy
pandas
scipy
scikit-learn
opencv-python-headless
pillow
matplotlib
pyyaml
pytest